+++ Libplanet.Tests/Net/SwarmTest.AppProtocolVersion.cs
@@ -1,3 +1,5 @@
+using System.Collections.Concurrent;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Libplanet.Crypto;
@@ -58,19 +60,19 @@ namespace Libplanet.Tests.Net
         {
             var isCalled = false;
 
-            void GameHandler(object sender, DifferentProtocolVersionEventArgs e)
-            {
-                isCalled = true;
-            }
-
             var signer = new PrivateKey();
+            var v1 = new AppProtocolVersion(signer, 1);
             var v2 = new AppProtocolVersion(signer, 2);
-            var v3 = new AppProtocolVersion(signer, 3);
             var a = CreateSwarm(
                 _blockchains[0],
-                appProtocolVersion: v2,
-                differentVersionPeerEncountered: GameHandler);
-            var b = CreateSwarm(_blockchains[1], appProtocolVersion: v3);
+                appProtocolVersion: v1,
+                differentAppProtocolVersionEncountered: (_, ver, __) =>
+                {
+                    isCalled = true;
+                    return false;
+                }
+            );
+            var b = CreateSwarm(_blockchains[1], appProtocolVersion: v2);
 
             try
             {
@@ -87,6 +89,119 @@ namespace Libplanet.Tests.Net
 
                 a.Dispose();
                 b.Dispose();
+            }
+        }
+
+        [Fact(Timeout = Timeout)]
+        public async Task IgnoreUntrustedAppProtocolVersion()
+        {
+            var blockChain = _blockchains[0];
+
+            var signer = new PrivateKey();
+            var older = new AppProtocolVersion(signer, 2);
+            var newer = new AppProtocolVersion(signer, 3);
+
+            var untrustedSigner = new PrivateKey();
+            var untrustedOlder = new AppProtocolVersion(untrustedSigner, 2);
+            var untrustedNewer = new AppProtocolVersion(untrustedSigner, 3);
+
+            _output.WriteLine("Trusted version signer: {0}", signer.PublicKey.ToAddress());
+            _output.WriteLine(
+                "Untrusted version signer: {0}",
+                untrustedSigner.PublicKey.ToAddress()
+            );
+
+            var logs = new ConcurrentDictionary<Peer, AppProtocolVersion>();
+
+            bool DifferentAppProtocolVersionEncountered(
+                Peer peer,
+                AppProtocolVersion peerVersion,
+                AppProtocolVersion localVersion
+            )
+            {
+                logs[peer] = peerVersion;
+                return false;
+            }
+
+            var a = CreateSwarm(
+                blockChain,
+                appProtocolVersion: older,
+                trustedAppProtocolVersionSigners: new[] { signer.PublicKey },
+                differentAppProtocolVersionEncountered: DifferentAppProtocolVersionEncountered
+            );
+            var b = CreateSwarm(
+                blockChain,
+                appProtocolVersion: newer,
+                trustedAppProtocolVersionSigners: new[] { signer.PublicKey }
+            );
+            var c = CreateSwarm(
+                blockChain,
+                appProtocolVersion: older,
+                trustedAppProtocolVersionSigners: new[] { signer.PublicKey }
+            );
+            var d = CreateSwarm(
+                blockChain,
+                appProtocolVersion: newer,
+                trustedAppProtocolVersionSigners: new[] { signer.PublicKey }
+            );
+            var e = CreateSwarm(
+                blockChain,
+                appProtocolVersion: untrustedOlder,
+                trustedAppProtocolVersionSigners: new[] { untrustedSigner.PublicKey }
+            );
+            var f = CreateSwarm(
+                blockChain,
+                appProtocolVersion: untrustedNewer,
+                trustedAppProtocolVersionSigners: new[] { untrustedSigner.PublicKey }
+            );
+
+            try
+            {
+                await StartAsync(c);
+                await StartAsync(d);
+                await StartAsync(e);
+                await StartAsync(f);
+
+                var peers = new[] { c.AsPeer, d.AsPeer, e.AsPeer, f.AsPeer };
+
+                foreach (var peer in peers)
+                {
+                    await a.AddPeersAsync(new[] { peer }, null);
+                    await b.AddPeersAsync(new[] { peer }, null);
+                }
+
+                Assert.Equal(new[] { c.AsPeer }, a.Peers.ToArray());
+                Assert.Equal(new[] { d.AsPeer }, b.Peers.ToArray());
+
+                _output.WriteLine("Logged encountered peers:");
+                foreach (KeyValuePair<Peer, AppProtocolVersion> kv in logs)
+                {
+                    _output.WriteLine(
+                        "{0}; {1}; {2} -> {3}",
+                        kv.Key,
+                        kv.Value.Version,
+                        kv.Value.Signer,
+                        kv.Value.Verify(signer.PublicKey) ? "verified" : "not verified"
+                    );
+                }
+
+                Assert.Equal(
+                    new Dictionary<Peer, AppProtocolVersion>
+                    {
+                        [d.AsPeer] = d.AsPeer.AppProtocolVersion,
+                    },
+                    logs
+                );
+            }
+            finally
+            {
+                await StopAsync(c);
+                await StopAsync(d);
+
+                a.Dispose();
+                b.Dispose();
+                c.Dispose();
+                d.Dispose();
             }
         }
     }