+++ Libplanet/Net/Peer.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics.Contracts;
+using System.Linq;
 using System.Net;
 using System.Runtime.Serialization;
 using Bencodex;
@@ -68,7 +70,6 @@ namespace Libplanet.Net
         /// <summary>
         /// The corresponding application protocol version of this peer.
         /// </summary>
-        /// <seealso cref="NetMQTransport.DifferentVersionPeerEncountered"/>
         [IgnoreDuringEquals]
         [Pure]
         public AppProtocolVersion AppProtocolVersion { get; }
@@ -83,6 +84,41 @@ namespace Libplanet.Net
 
         [Pure]
         public IPAddress PublicIPAddress { get; }
+
+        /// <summary>
+        /// Determines if the peer is compatible with the given criteria.
+        /// </summary>
+        /// <param name="localVersion">The version of the currently running application.</param>
+        /// <param name="trustedSigners">A list of signers to trust.  <em>An empty list means
+        /// to trust no one, so that this method always returns <c>false</c>.</em>  Conversely,
+        /// <c>null</c> means to <em>trust anyone</em>.  Be aware of the difference of those two.
+        /// </param>
+        /// <param name="recognizer">A callback to determine an encountered peer's
+        /// <see cref="AppProtocolVersion"/> is compatible with the currently running
+        /// application, in the case of the peer's version signature is trustworthy but its
+        /// version number is not the same.</param>
+        /// <returns><c>true</c> if the peer is compatible with the given criteria.
+        /// Otherwise <c>false</c>.</returns>
+        public bool IsCompatibleWith(
+            AppProtocolVersion localVersion,
+            IEnumerable<PublicKey> trustedSigners = null,
+            DifferentAppProtocolVersionEncountered recognizer = null
+        )
+        {
+            if (AppProtocolVersion.Equals(localVersion))
+            {
+                return true;
+            }
+
+            if (trustedSigners is null || trustedSigners.Any(s => AppProtocolVersion.Verify(s)))
+            {
+                return recognizer is null
+                    ? false
+                    : recognizer(this, AppProtocolVersion, localVersion);
+            }
+
+            return false;
+        }
 
         /// <inheritdoc/>
         public virtual void GetObjectData(SerializationInfo info, StreamingContext context)