+++ Libplanet/Net/Swarm.cs
@@ -66,8 +66,8 @@ namespace Libplanet.Net
             int? listenPort = null,
             DateTimeOffset? createdAt = null,
             IEnumerable<IceServer> iceServers = null,
-            EventHandler<DifferentProtocolVersionEventArgs>
-                differentVersionPeerEncountered = null)
+            DifferentAppProtocolVersionEncountered differentAppProtocolVersionEncountered = null,
+            IEnumerable<PublicKey> trustedAppProtocolVersionSigners = null)
             : this(
                 blockChain,
                 privateKey,
@@ -79,7 +79,8 @@ namespace Libplanet.Net
                 listenPort,
                 createdAt,
                 iceServers,
-                differentVersionPeerEncountered)
+                differentAppProtocolVersionEncountered,
+                trustedAppProtocolVersionSigners)
         {
         }
 
@@ -94,7 +95,8 @@ namespace Libplanet.Net
             int? listenPort = null,
             DateTimeOffset? createdAt = null,
             IEnumerable<IceServer> iceServers = null,
-            EventHandler<DifferentProtocolVersionEventArgs> differentVersionPeerEncountered = null)
+            DifferentAppProtocolVersionEncountered differentAppProtocolVersionEncountered = null,
+            IEnumerable<PublicKey> trustedAppProtocolVersionSigners = null)
         {
             BlockChain = blockChain ?? throw new ArgumentNullException(nameof(blockChain));
             _store = BlockChain.Store;
@@ -114,6 +116,8 @@ namespace Libplanet.Net
             _runningMutex = new AsyncLock();
 
             _appProtocolVersion = appProtocolVersion;
+            TrustedAppProtocolVersionSigners =
+                trustedAppProtocolVersionSigners?.ToImmutableHashSet();
 
             string loggerId = _privateKey.PublicKey.ToAddress().ToHex();
             _logger = Log.ForContext<Swarm<T>>()
@@ -122,13 +126,14 @@ namespace Libplanet.Net
             Transport = new NetMQTransport(
                 _privateKey,
                 _appProtocolVersion,
+                TrustedAppProtocolVersionSigners,
                 tableSize,
                 bucketSize,
                 workers,
                 host,
                 listenPort,
                 iceServers,
-                differentVersionPeerEncountered,
+                differentAppProtocolVersionEncountered,
                 ProcessMessageHandler,
                 _logger);
         }
@@ -167,6 +172,8 @@ namespace Libplanet.Net
         /// synchronizes with.
         /// </summary>
         public BlockChain<T> BlockChain { get; private set; }
+
+        public IImmutableSet<PublicKey> TrustedAppProtocolVersionSigners { get; }
 
         internal ITransport Transport { get; private set; }